requirements.txt
aiohttp[speedups]
aiogram
python-decouple
ujson
# uvloop